chData, chInfo 스키마 변경 관련 코드 수정

+++ retriever/src/main/java/com/team7/retriever/entity/ChatBots.java
@@ -13,7 +13,7 @@ import java.util.List;
 public class ChatBots {
     @Id
     private String id;
-    private int channelId;
+    private long channelId;
     private LocalDateTime updatedAt;
     private List<Integer> chatIds = new ArrayList<>();
 }
